requirements.txt
torch
torchvision
torchaudio
cython
matplotlib
numpy
pandas
python-dotenv
ultralytics
opencv
pyyaml
ninja
neo4j
neo4j-driver
langchain
langchain-community
langchain-openai
langgraph
openai
tqdm
requests
Pillow
openmim
mmcv==2.1.0
mmengine
mmdet
mmpose
mmaction2
 